Improve cartridge, nextevent and *support: only JugSupport is now required by JugController
