+++ montpellier-jug-wisdom/src/main/java/org/jug/montpellier/core/api/JugSupport.java
@@ -1,0 +1,9 @@
+package org.jug.montpellier.core.api;
+
+
+import java.util.List;
+
+
+public interface JugSupport extends CartridgeSupport, NextEventSupport, PartnerSupport {
+
+}
