+++ montpellier-jug-wisdom/src/main/resources/templates/admin.mst.html
@@ -10,7 +10,9 @@
             <a href="#" class="list-group-item">Other</a>
         </div>
     </div>
-    {{{propertysheet}}}
+    <div class="col-md-9 propertysheet">
+      {{{propertysheet}}}
+    </div>
 </div>
