+++ montpellier-jug-wisdom/src/main/resources/templates/editors/propertysheet.mst.html
@@ -1,28 +1,26 @@
-<div class="col-md-6 propertysheet">
-    <div class="panel panel-default">
-        <form name="propertysheet" class="form-horizontal" method="post">
-            <div class="panel-heading">Edit News</div>
-            <div class="panel-body">
+<div class="panel panel-default">
+    <form name="propertysheet" class="form-horizontal" method="post">
+        <div class="panel-heading">Edit News</div>
+        <div class="panel-body">
 
-                {{#properties}}
-                <div class="{{editorName}} form-group {{^visible}}hidden{{/visible}}">
-                    <div class="row">
-                        <label class="col-sm-4  control-label" for="{{name}}">{{displayname}}</label>
-                        <div class="col-sm-6">{{{editor}}}</div>
-                    </div>
-                    <div class="row">
-                        <span class="col-sm-4"></span>
-                        <div class="col-sm-6 small">{{description}}</div>
-                    </div>
+            {{#properties}}
+            <div class="{{editorName}} form-group {{^visible}}hidden{{/visible}}">
+                <div class="row">
+                    <label class="col-sm-4  control-label" for="{{name}}">{{displayname}}</label>
+                    <div class="col-sm-6">{{{editor}}}</div>
                 </div>
-                {{/properties}}
+                <div class="row">
+                    <span class="col-sm-4"></span>
+                    <div class="col-sm-6 small">{{description}}</div>
+                </div>
+            </div>
+            {{/properties}}
 
-            </div>
-            <div class="panel-footer">
-                <button class="btn btn-primary" type="submit">Save</button>
-                <button class="btn btn-default" type="button">Cancel</button>
-            </div>
-        </form>
-    </div>
+        </div>
+        <div class="panel-footer">
+            <button class="btn btn-primary" type="submit">Save</button>
+            <button class="btn btn-default" type="button">Cancel</button>
+        </div>
+    </form>
 </div>
-<div class="col-md-2"></div>
+
