+++ montpellier-jug-wisdom/src/main/resources/templates/events.mst.html
@@ -1,4 +1,7 @@
 {{> layouts/header}}
 
+{{#nextEvent}}
+    {{> layouts/nextEventView}}
+{{/nextEvent}}
 
 {{> layouts/footer}}
