--- montpellier-jug-wisdom/src/main/resources/templates/form.mst.html
@@ -1,15 +1,0 @@
-{{> layouts/header}}
-
-<link rel="stylesheet" href="/assets/form.css" />
-
-<div class="form-container">
-    <form class="form" action="{{urlSubmit}}" method="POST">
-        {{#properties}}
-        <div class="form-item">
-            <div class="form-label">{{name}} : </div>
-            <input type="text" name="{{name}}" value="{{value}}" />
-        </div>
-        {{/properties}}
-        <input type="submit" value="Ok">
-    </form>
-</div>
