+++ montpellier-jug-wisdom/src/main/resources/templates/layouts/header.mst.html
@@ -4,8 +4,8 @@
     <meta charset="utf-8"/>
     <meta http-equiv="X-UA-Compatible" content="IE=edge"/>
     <meta name="viewport" content="width=device-width, initial-scale=1"/>
-    <meta name="description" content="Wisdom Framework Monitoring"/>
-    <meta name="author" content="The Wisdom Framework Team"/>
+    <meta name="description" content="The JUG Montpellier Web Site"/>
+    <meta name="author" content="The JUG Montpellier Team with love"/>
     <link rel="shortcut icon" href="/assets/images/owl-small.png"/>
 
     <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.2/css/bootstrap.min.css">
@@ -24,6 +24,3 @@
             {{/cartridges}}
         </ul>
     </nav>
-    {{#nextEvent}}
-        {{> nextEventView}}
-    {{/nextEvent}}
