--- montpellier-jug-wisdom/src/main/resources/templates/welcome.mst.html
@@ -1,5 +1,0 @@
-{{> layouts/header}}
-
-Welcome to this the new brand website!
-
-{{> layouts/footer}}
